начало нескольких секундомеров
