+++ app/src/main/java/com/gb/stopwatch/view/MainActivity.kt
@@ -5,7 +5,6 @@ import android.widget.Button
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import com.gb.stopwatch.R
-import com.gb.stopwatch.model.time.*
 import com.gb.stopwatch.viewmodel.StopwatchListOrchestrator
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -15,21 +14,8 @@ import kotlinx.coroutines.launch
 
 class MainActivity : AppCompatActivity() {
 
-    private val timestampProvider = object : TimestampProvider {
-        override fun getMilliseconds(): Long {
-            return System.currentTimeMillis()
-        }
-    }
     private val stopwatchListOrchestrator = StopwatchListOrchestrator(
-        StopwatchStateHolder(
-            StopwatchStateCalculator(
-                timestampProvider,
-                ElapsedTimeCalculator(timestampProvider)
-            ),
-            ElapsedTimeCalculator(timestampProvider),
-            TimestampMillisecondsFormatter()
-        ),
-        CoroutineScope(Dispatchers.Main + SupervisorJob())
+        scope = CoroutineScope(Dispatchers.Main + SupervisorJob())
     )
 
     override fun onCreate(savedInstanceState: Bundle?) {
