+++ app/src/main/res/layout/stopwatch_item.xml
